requirements.txt
Flask==2.3.3
python-docx==0.8.11
PyPDF2==3.0.1
openpyxl==3.1.2
Werkzeug==2.3.7
Jinja2==3.1.2
click==8.1.7
itsdangerous==2.1.2
MarkupSafe==2.1.3
python-magic==0.4.27
chardet==5.2.0
fonttools==4.42.1
Pillow==10.0.0
pandas==2.1.1
numpy==1.25.2